Task 7

+++ 0x00-python_variable_annotations/7-to_kv.py
@@ -5,7 +5,7 @@ Complex types - string and int/float to tuple
 from typing import Union, Tuple
 
 
-def to_kv(k: str, v: Union[float, int]) -> Tuple[str, float]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
         """
         Return str or float
         """
